test : PathSteps 의 경로조회 메서드 문서 테스트에서 사용하도록 변경
- PathDocumentationSteps 제거
- DocumentSteps 추가 (given 메서드 추가)
- PathDocumentation 에서 PathSteps 요청 메서드 사용 (문서화에 사용될 given 파라미터로 넘겨줌)
